// toinpark-backend/src/common/enums/SystemSettings.ts
export enum EnumSystemSetting {
  TOIN_CONVENTION_RATE = 'toinConventionRate',
  FAVICON = 'favicon',
  LOGO = 'logo',
  WHATSAPP = 'whatsApp',
  ORGANIZATION_NAME = 'organizationName',
  ORGANIZATION_EMAIL = 'organizationEmail',
  TELEGRAM = 'telegram',
  SITE_TITLE = 'siteTitle',
  AIRDROP_EVENT_MESSAGE = 'airdropEventMessage',  
  ENTRY_BONUS_TOIN = 'entryBonusToin',
  KYC_BONUS_TOINS = 'kycBonusToin',
  SOCIAL_BONUS_TOINS = 'socialBonusToin',
  PDF = 'pdf',
  FACEBOOK_URL = 'facebookUrl',
  YOUTUBE_URL = 'youtubeUrl',
  LINKEDIN_URL = 'linkedinUrl',
  INSTAGRAM_URL = 'instagramUrl',
  MIN_USDT_WITHDRAWAL_AMOUNT = 'minUSDTWithdrawalAmount',
  PLATFORM_WITHDRAWAL_FEE_PERCENTAGE = 'platformWithdrawalFeePercentage'
}